escape html on upload

+++ server/src/lib/routes/api/data/jobs.ts
@@ -1,6 +1,7 @@
 import { Router, Request, Response } from "express";
 import Job from "../../../models/jobs";
 import passport from "passport";
+import validator from "validator";
 
 const router = Router();
 
@@ -18,6 +19,9 @@ router.get("/jobs/:id", async (req: Request, res: Response) => {
 });
 
 router.post("/jobs", passport.authenticate("local"), async (req: Request, res: Response) => {
+    Object.entries(req.body).forEach(([key, value]) => {
+        req.body[key] = validator.escape(value as string);
+    });
     const job = await Job.create(req.body);
     await job.save();
     res.sendStatus(200);
@@ -25,14 +29,23 @@ router.post("/jobs", passport.authenticate("local"), async (req: Request, res: R
 
 router.put("/jobs/:id", passport.authenticate("local"), async (req: Request, res: Response) => {
     const { id } = req.query;
-    await Job.findByIdAndUpdate(id, req.body);
-    res.sendStatus(200);
+    if (id) {
+        Object.entries(req.body).forEach(([key, value]) => {
+            req.body[key] = validator.escape(value as string);
+        });
+        await Job.findByIdAndUpdate(id, req.body);
+        res.sendStatus(200);
+    }
+    res.sendStatus(400);
 });
 
 router.delete("/jobs/:id", passport.authenticate("local"), async (req: Request, res: Response) => {
     const { id } = req.query;
-    await Job.findByIdAndDelete(id);
-    res.sendStatus(200);
+    if (id) {
+        await Job.findByIdAndDelete(id);
+        res.sendStatus(200);
+    }
+    res.send(400);
 });
 
 export default router;
