+++ server/src/lib/routes/api/data/projects.ts
@@ -1,6 +1,7 @@
 import { Router, Request, Response } from "express";
 import passport from "passport";
 import Project from "../../../models/projects";
+import validator from "validator";
 
 const router = Router();
 
@@ -18,6 +19,9 @@ router.get("/projects/:id", async (req: Request, res: Response) => {
 });
 
 router.post("/projects", passport.authenticate("local"), async (req: Request, res: Response) => {
+    Object.entries(req.body).forEach(([key, value]) => {
+        req.body[key] = validator.escape(value as string);
+    });
     const project = await Project.create(req.body);
     await project.save();
     res.sendStatus(200);
@@ -25,14 +29,28 @@ router.post("/projects", passport.authenticate("local"), async (req: Request, re
 
 router.put("/projects/:id", passport.authenticate("local"), async (req: Request, res: Response) => {
     const { id } = req.query;
-    await Project.findByIdAndUpdate(id, req.body);
-    res.sendStatus(200);
+
+    if (id) {
+        Object.entries(req.body).forEach(([key, value]) => {
+            req.body[key] = validator.escape(value as string);
+        });
+        await Project.findByIdAndUpdate(id, req.body);
+        res.sendStatus(200);
+    }
+    res.sendStatus(400);
 });
 
 router.delete("/projects/id", passport.authenticate("local"), async (req: Request, res: Response) => {
     const { id } = req.query;
-    await Project.findByIdAndDelete(id);
-    res.sendStatus(200);
+
+    if (id) {
+        Object.entries(req.body).forEach(([key, value]) => {
+            req.body[key] = validator.escape(value as string);
+        });
+        await Project.findByIdAndDelete(id);
+        res.sendStatus(200);
+    }
+    res.sendStatus(400);
 });
 
 export default router;
